Update and Rename to DistinctNumbers.java
Write a program that reads in 10 numbers and 
displays the number of distinct numbers and 
the distinct numbers in their input order and 
separated by exactly one space (i.e., if a 
number appears multiple times, it is displayed 
only once).  (Hint:  Read a number and store it 
in an array if it is new.  If the number is 
already in the array, ignore it.)  After the 
input, the array contains the distinct numbers.